+++ config/appearance.php
@@ -8,10 +8,10 @@
  * @link    https://www.studiopress.com/
  */
 
-$genesis_sample_default_colors = array(
+$genesis_sample_default_colors = [
 	'link'   => '#0073e5',
 	'accent' => '#0073e5',
-);
+];
 
 $genesis_sample_link_color = get_theme_mod(
 	'genesis_sample_link_color',
@@ -26,7 +26,7 @@ $genesis_sample_accent_color = get_theme_mod(
 $genesis_sample_link_color_contrast   = genesis_sample_color_contrast( $genesis_sample_link_color );
 $genesis_sample_link_color_brightness = genesis_sample_color_brightness( $genesis_sample_link_color, 35 );
 
-return array(
+return [
 	'fonts-url'            => 'https://fonts.googleapis.com/css?family=Source+Sans+Pro:400,400i,600,700&display=swap',
 	'content-width'        => 1062,
 	'button-bg'            => $genesis_sample_link_color,
@@ -34,38 +34,38 @@ return array(
 	'button-outline-hover' => $genesis_sample_link_color_brightness,
 	'link-color'           => $genesis_sample_link_color,
 	'default-colors'       => $genesis_sample_default_colors,
-	'editor-color-palette' => array(
-		array(
+	'editor-color-palette' => [
+		[
 			'name'  => __( 'Custom color', 'genesis-sample' ), // Called “Link Color” in the Customizer options. Renamed because “Link Color” implies it can only be used for links.
 			'slug'  => 'theme-primary',
 			'color' => $genesis_sample_link_color,
-		),
-		array(
+		],
+		[
 			'name'  => __( 'Accent color', 'genesis-sample' ),
 			'slug'  => 'theme-secondary',
 			'color' => $genesis_sample_accent_color,
-		),
-	),
-	'editor-font-sizes'    => array(
-		array(
+		],
+	],
+	'editor-font-sizes'    => [
+		[
 			'name' => __( 'Small', 'genesis-sample' ),
 			'size' => 12,
 			'slug' => 'small',
-		),
-		array(
+		],
+		[
 			'name' => __( 'Normal', 'genesis-sample' ),
 			'size' => 18,
 			'slug' => 'normal',
-		),
-		array(
+		],
+		[
 			'name' => __( 'Large', 'genesis-sample' ),
 			'size' => 20,
 			'slug' => 'large',
-		),
-		array(
+		],
+		[
 			'name' => __( 'Larger', 'genesis-sample' ),
 			'size' => 24,
 			'slug' => 'larger',
-		),
-	),
-);
+		],
+	],
+];