+++ config/child-theme-settings.php
@@ -10,8 +10,8 @@
  * @link    https://www.studiopress.com/
  */
 
-return array(
-	GENESIS_SETTINGS_FIELD => array(
+return [
+	GENESIS_SETTINGS_FIELD => [
 		'blog_cat_num'              => 6,
 		'breadcrumb_home'           => 0,
 		'breadcrumb_front_page'     => 0,
@@ -24,10 +24,10 @@ return array(
 		'content_archive'           => 'full',
 		'content_archive_limit'     => 0,
 		'content_archive_thumbnail' => 0,
-		'image_size'                => '',
-		'image_alignment'           => 'alignleft',
+		'image_size'                => 'genesis-singular-images',
+		'image_alignment'           => 'aligncenter',
 		'posts_nav'                 => 'numeric',
 		'site_layout'               => 'content-sidebar',
-	),
+	],
 	'posts_per_page'       => 6,
-);
+];