+++ config/onboarding-shared.php
@@ -1,0 +1,139 @@
+<?php
+/**
+ * Genesis Sample.
+ *
+ * Onboarding config shared between Starter Packs.
+ *
+ * Genesis Starter Packs give you a choice of content variation when activating
+ * the theme. The content below is common to all packs for this theme.
+ *
+ * @package Genesis Sample
+ * @author  StudioPress
+ * @license GPL-2.0-or-later
+ * @link    https://www.studiopress.com/
+ */
+
+return [
+	'plugins'          => [
+		[
+			'name'       => __( 'Atomic Blocks', 'genesis-sample' ),
+			'slug'       => 'atomic-blocks/atomicblocks.php',
+			'public_url' => 'https://atomicblocks.com/',
+		],
+		[
+			'name'       => __( 'Simple Social Icons', 'genesis-sample' ),
+			'slug'       => 'simple-social-icons/simple-social-icons.php',
+			'public_url' => 'https://wordpress.org/plugins/simple-social-icons/',
+		],
+		[
+			'name'       => __( 'Genesis eNews Extended (Third Party)', 'genesis-sample' ),
+			'slug'       => 'genesis-enews-extended/plugin.php',
+			'public_url' => 'https://wordpress.org/plugins/genesis-enews-extended/',
+		],
+		[
+			'name'       => __( 'WPForms Lite (Third Party)', 'genesis-sample' ),
+			'slug'       => 'wpforms-lite/wpforms.php',
+			'public_url' => 'https://wordpress.org/plugins/wpforms-lite/',
+		],
+	],
+	'content'          => [
+		'blocks'  => [
+			'post_title'     => 'Block Content Examples',
+			'post_content'   => require dirname( __FILE__ ) . '/import/content/block-examples.php',
+			'post_type'      => 'page',
+			'post_status'    => 'publish',
+			'comment_status' => 'closed',
+			'ping_status'    => 'closed',
+			'meta_input'     => [ '_genesis_layout' => 'full-width-content' ],
+		],
+		'about'   => [
+			'post_title'     => 'About Us',
+			'post_content'   => require dirname( __FILE__ ) . '/import/content/about.php',
+			'post_type'      => 'page',
+			'post_status'    => 'publish',
+			'featured_image' => CHILD_URL . '/config/import/images/about.jpg',
+			'comment_status' => 'closed',
+			'ping_status'    => 'closed',
+			'meta_input'     => [
+				'_genesis_layout'              => 'full-width-content',
+				'_genesis_hide_singular_image' => true,
+			],
+		],
+		'contact' => [
+			'post_title'     => 'Contact Us',
+			'post_content'   => require dirname( __FILE__ ) . '/import/content/contact.php',
+			'post_type'      => 'page',
+			'post_status'    => 'publish',
+			'comment_status' => 'closed',
+			'ping_status'    => 'closed',
+		],
+		'landing' => [
+			'post_title'     => 'Landing Page',
+			'post_content'   => require dirname( __FILE__ ) . '/import/content/landing-page.php',
+			'post_type'      => 'page',
+			'post_status'    => 'publish',
+			'page_template'  => 'page-templates/landing.php',
+			'comment_status' => 'closed',
+			'ping_status'    => 'closed',
+			'meta_input'     => [
+				'_genesis_layout'              => 'full-width-content',
+				'_genesis_hide_breadcrumbs'    => true,
+				'_genesis_hide_singular_image' => true,
+			],
+		],
+	],
+	'navigation_menus' => [
+		'primary' => [
+			'homepage' => [
+				'title' => 'Home',
+			],
+			'about'    => [
+				'title' => 'About Us',
+			],
+			'contact'  => [
+				'title' => 'Contact Us',
+			],
+			'blocks'   => [
+				'title' => 'Block Examples',
+			],
+			'landing'  => [
+				'title' => 'Landing Page',
+			],
+		],
+	],
+	'widgets'          => [
+		'footer-1' => [
+			[
+				'type' => 'text',
+				'args' => [
+					'title'  => 'Design',
+					'text'   => '<p>With an emphasis on typography, white space, and mobile-optimized design, your website will look absolutely breathtaking.</p><p><a href="#">Learn more about design</a>.</p>',
+					'filter' => 1,
+					'visual' => 1,
+				],
+			],
+		],
+		'footer-2' => [
+			[
+				'type' => 'text',
+				'args' => [
+					'title'  => 'Content',
+					'text'   => '<p>Our team will teach you the art of writing audience-focused content that will help you achieve the success you truly deserve.</p><p><a href="#">Learn more about content</a>.</p>',
+					'filter' => 1,
+					'visual' => 1,
+				],
+			],
+		],
+		'footer-3' => [
+			[
+				'type' => 'text',
+				'args' => [
+					'title'  => 'Strategy',
+					'text'   => '<p>We help creative entrepreneurs build their digital business by focusing on three key elements of a successful online platform.</p><p><a href="#">Learn more about strategy</a>.</p>',
+					'filter' => 1,
+					'visual' => 1,
+				],
+			],
+		],
+	],
+];