+++ config/post-type-supports.php
@@ -2,22 +2,19 @@
 /**
  * Genesis Sample child theme.
  *
+ * Post type supports.
+ *
  * @package Genesis Sample
  * @author  StudioPress
  * @license GPL-2.0-or-later
  * @link    https://my.studiopress.com/themes/genesis-sample/
  */
 
-/**
- * Genesis responsive menus settings. (Requires Genesis 3.0+.)
- */
 return [
-	'script' => [
-		'menuClasses' => [
-			'others' => [ '.nav-primary' ],
-		],
+	'post' => [
+		'genesis-singular-images',
 	],
-	'extras' => [
-		'media_query_width' => '960px',
+	'page' => [
+		'genesis-singular-images',
 	],
 ];