+++ lib/customize.php
@@ -24,69 +24,89 @@ function genesis_sample_customizer_register( $wp_customize ) {
 
 	$wp_customize->add_setting(
 		'genesis_sample_link_color',
-		array(
+		[
 			'default'           => $appearance['default-colors']['link'],
 			'sanitize_callback' => 'sanitize_hex_color',
-		)
+		]
 	);
 
 	$wp_customize->add_control(
 		new WP_Customize_Color_Control(
 			$wp_customize,
 			'genesis_sample_link_color',
-			array(
+			[
 				'description' => __( 'Change the color of post info links and button blocks, the hover color of linked titles and menu items, and more.', 'genesis-sample' ),
 				'label'       => __( 'Link Color', 'genesis-sample' ),
 				'section'     => 'colors',
 				'settings'    => 'genesis_sample_link_color',
-			)
+			]
 		)
 	);
 
 	$wp_customize->add_setting(
 		'genesis_sample_accent_color',
-		array(
+		[
 			'default'           => $appearance['default-colors']['accent'],
 			'sanitize_callback' => 'sanitize_hex_color',
-		)
+		]
 	);
 
 	$wp_customize->add_control(
 		new WP_Customize_Color_Control(
 			$wp_customize,
 			'genesis_sample_accent_color',
-			array(
+			[
 				'description' => __( 'Change the default hover color for button links, menu buttons, and submit buttons. The button block uses the Link Color.', 'genesis-sample' ),
 				'label'       => __( 'Accent Color', 'genesis-sample' ),
 				'section'     => 'colors',
 				'settings'    => 'genesis_sample_accent_color',
-			)
+			]
 		)
 	);
 
 	$wp_customize->add_setting(
 		'genesis_sample_logo_width',
-		array(
+		[
 			'default'           => 350,
 			'sanitize_callback' => 'absint',
-		)
+			'validate_callback' => 'genesis_sample_validate_logo_width',
+		]
 	);
 
 	// Add a control for the logo size.
 	$wp_customize->add_control(
 		'genesis_sample_logo_width',
-		array(
+		[
 			'label'       => __( 'Logo Width', 'genesis-sample' ),
 			'description' => __( 'The maximum width of the logo in pixels.', 'genesis-sample' ),
 			'priority'    => 9,
 			'section'     => 'title_tagline',
 			'settings'    => 'genesis_sample_logo_width',
 			'type'        => 'number',
-			'input_attrs' => array(
+			'input_attrs' => [
 				'min' => 100,
-			),
+			],
 
-		)
+		]
 	);
+
+}
+
+/**
+ * Displays a message if the entered width is not numeric or greater than 100.
+ *
+ * @param object $validity The validity status.
+ * @param int    $width The width entered by the user.
+ * @return int The new width.
+ */
+function genesis_sample_validate_logo_width( $validity, $width ) {
+
+	if ( empty( $width ) || ! is_numeric( $width ) ) {
+		$validity->add( 'required', __( 'You must supply a valid number.', 'genesis-sample' ) );
+	} elseif ( $width < 100 ) {
+		$validity->add( 'logo_too_small', __( 'The logo width cannot be less than 100.', 'genesis-sample' ) );
+	}
+
+	return $validity;
 
 }