+++ lib/gutenberg/init.php
@@ -19,7 +19,7 @@ function genesis_sample_enqueue_gutenberg_frontend_styles() {
 	wp_enqueue_style(
 		genesis_get_theme_handle() . '-gutenberg',
 		get_stylesheet_directory_uri() . '/lib/gutenberg/front-end.css',
-		array( genesis_get_theme_handle() ),
+		[ genesis_get_theme_handle() ],
 		genesis_get_theme_version()
 	);
 
@@ -38,7 +38,7 @@ function genesis_sample_block_editor_styles() {
 	wp_enqueue_style(
 		genesis_get_theme_handle() . '-gutenberg-fonts',
 		$appearance['fonts-url'],
-		array(),
+		[],
 		genesis_get_theme_version()
 	);