+++ page-templates/landing.php
@@ -43,9 +43,6 @@ function genesis_sample_dequeue_skip_links() {
 
 }
 
-// Forces full width content layout.
-add_filter( 'genesis_site_layout', '__genesis_return_full_width_content' );
-
 // Removes site header elements.
 remove_action( 'genesis_header', 'genesis_header_markup_open', 5 );
 remove_action( 'genesis_header', 'genesis_do_header' );
@@ -53,9 +50,6 @@ remove_action( 'genesis_header', 'genesis_header_markup_close', 15 );
 
 // Removes navigation.
 remove_theme_support( 'genesis-menus' );
-
-// Removes breadcrumbs.
-remove_action( 'genesis_before_loop', 'genesis_do_breadcrumbs' );
 
 // Removes footer widgets.
 remove_action( 'genesis_before_footer', 'genesis_footer_widget_areas' );